Refactoring, marking
