+++ CryptoMarkets/ViewControllers/ExchangesListViewController.swift
@@ -8,38 +8,23 @@
 import UIKit
 
 class ExchangesListViewController: UITableViewController {
-    let dataManager = NetworkManager.shared
-    var exchanges: [Exchange] = []
-
+    // MARK: - IBOutlets
     @IBOutlet weak var activityIndicator: UIActivityIndicatorView!
     
+    // MARK: - Private properties
+    private var exchanges: [Exchange] = []
+    
+    // MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
-        dataManager.fetchAllExchanges { result in
-                   switch result {
-       
-                   case .success(let exchanges):
-                       self.exchanges = exchanges
-                       DispatchQueue.main.async {
-                           self.activityIndicator.isHidden = true
-                           self.tableView.reloadData()
-                       }
- 
-                       print(exchanges)
-              
-                   case .failure(let error):
-                       print(error)
-                   }
-               }
+        fetchAllChanges()
     }
-
+    
     // MARK: - Table view data source
-
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         exchanges.count
     }
-
-
+    
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "exchangeCell", for: indexPath)
         var content = cell.defaultContentConfiguration()
@@ -48,16 +33,28 @@ class ExchangesListViewController: UITableViewController {
         cell.contentConfiguration = content
         return cell
     }
-
+    
     // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         guard let detailedVC = segue.destination as? DetailedExchangeViewController else { return }
         guard let index = tableView.indexPathForSelectedRow  else { return }
         let selectedExchange = exchanges[index.row]
         detailedVC.exchange = selectedExchange
     }
-
-
+    
+    // MARK: - Private methods
+    private func fetchAllChanges() {
+        NetworkManager.shared.fetchAllExchanges { result in
+            switch result {
+            case .success(let exchanges):
+                self.exchanges = exchanges
+                DispatchQueue.main.async {
+                    self.activityIndicator.stopAnimating()
+                    self.tableView.reloadData()
+                }
+            case .failure(let error):
+                print(error)
+            }
+        }
+    }
 }
