+++ CryptoMarkets/ViewControllers/MarketsByCurrencyViewController.swift
@@ -8,52 +8,50 @@
 import UIKit
 
 class MarketsByCurrencyViewController: UITableViewController {
+    // MARK: - IBOutlets
+    @IBOutlet weak var activityIndicator: UIActivityIndicatorView!
+    
+    // MARK: - Public properties
     var currency: Currency!
     
+    // MARK: - Private properties
     private var markets: [Market] = []
-    private let dataManager = NetworkManager.shared
-
+    
+    // MARK: - Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
         title = currency.name
-        
-        dataManager.fetchMarketsForCurrencyBy(id: currency.asset_id) { result in
-            switch result {
-    
-            case .success(let markets):
-                self.markets = markets
-                DispatchQueue.main.async {
-                    self.tableView.reloadData()
-                }
-            case .failure(let error):
-                print(error.localizedDescription)
-            }
-        }
-
-        // Uncomment the following line to preserve selection between presentations
-        // self.clearsSelectionOnViewWillAppear = false
-
-        // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
-        // self.navigationItem.rightBarButtonItem = self.editButtonItem
+        fetchMarketsForCurrency()
     }
 
     // MARK: - Table view data source
-
-
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
         return markets.count
     }
-
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "marketCurrencyCell", for: indexPath) as! CurrencyForMarketTableViewCell
 
         let market = markets[indexPath.row]
         cell.configureCellFor(market)
-        // Configure the cell...
-
+        
         return cell
+    }
+    
+    // MARK: - Private methods
+    private func fetchMarketsForCurrency() {
+        NetworkManager.shared.fetchMarketsForCurrencyBy(id: currency.asset_id) { result in
+            switch result {
+            case .success(let markets):
+                self.markets = markets
+                DispatchQueue.main.async {
+                    self.activityIndicator.stopAnimating()
+                    self.tableView.reloadData()
+                }
+            case .failure(let error):
+                print(error.localizedDescription)
+            }
+        }
     }
 
 }
