--- CryptoMarkets/ViewControllers/StartViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  CryptoMarkets
-//
-//  Created by Nadzeya Shpakouskaya on 24/09/2021.
-//
-
-import UIKit
-
-class StartViewController: UIViewController {
-
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-    
-}
-
-
